# Excluding enrollmentComplete experiments from getValidExperiments query (#909)