This PR contains changes to improve the PROD latency issues due to enrolmentComplete experiments for post-experiment rule set to default condition. We are now excluding enrollmentComplete experiments from getValidExperiments and getValidExperimentsWithPreview queries.

This gives an optimization of 62% in terms of average api response time for 600 concurrent users. With Splitting of queries into 3 and Caching, the time is reduced by 91% in terms of average api response time for 600 concurrent users. Please find the detailed Load Testing Analysis in this [document](https://docs.google.com/document/d/1hsQpcF3ptcYaHfUaw_iqPrCv0AK4hx4YVJrdfYclNlM/edit?usp=sharing). Here is the Response Time Comparison table:
<meta charset="utf-8"><b style="font-weight:normal;" id="docs-internal-guid-a12e0a09-7fff-7017-ecb6-b4e70c70c1ba"><div dir="ltr" style="margin-left:0pt;" align="center">

Test Case | Concurrent Users | Trial 1 (Median Response Time) | Trial 2 (Median Response Time) | Avg of (Median Response Time)
-- | -- | -- | -- | --
No enrollmentComplete experiments (Total 8 experiments) | 600 | 5700 ms | 4300 ms | 5000 ms
All post-exp rule default experiments set to enrollmentComplete (5 out of 8 experiments) | 600 | 7300 ms | 7800 ms | 7500 ms
Exclude enrollmentComplete experiments from getValidExperiments(5 out of 8 experiments) | 600 | 3100 ms | 2600 ms | 2850 ms
Split queries into 3 + Exclude enrollmentComplete experiments from getValidExperiments(5 out of 8 experiments) | 600 | 2300 ms | 1800 ms | 2050 ms
Split queries into 3 + Caching + Exclude enrollmentComplete experiments from getValidExperiments(5 out of 8 experiments) | 600 | 850 ms | 550 ms | 700 ms

</div></b>

Outcome (For 600 concurrent users): 

- So, there is a 62% improvement when the time is reduced from 7500 ms to 2850 ms for excluding enrollmentComplete experiments from getValidExperiments

- So, there is a 91% improvement when the time is reduced from 7500 ms to 700 ms for (Split queries into 3 + Caching + excluding enrollmentComplete experiments) from getValidExperiments


<meta charset="utf-8"><b style="font-weight:normal;" id="docs-internal-guid-87030005-7fff-5804-3917-b12cb6662fdc"><p dir="ltr" style="line-height:1.38;margin-top:0pt;margin-bottom:0pt;"><span style="font-size:11pt;font-family:Arial,sans-serif;color:#000000;background-color:transparent;font-weight:700;font-style:normal;font-variant:normal;text-decoration:none;vertical-align:baseline;white-space:pre;white-space:pre-wrap;">Observation: Due to load on my machine, I tried with lesser concurrent users (200/400) as well, below are the results:</span></p><br /><div dir="ltr" style="margin-left:0pt;" align="center">

Test Case | Concurrent Users | Trail 1 (Median Response Time) | Trail 2 (Median Response Time) | Avg of (Median Response Time)
-- | -- | -- | -- | --
Split queries into 3 + Caching + Exclude enrollmentComplete experiments from getValidExperiments(5 out of 8 experiments) | 400 | 88 ms | 91 ms | 89.5 ms
Split queries into 3 + Caching + Exclude enrollmentComplete experiments from getValidExperiments(5 out of 8 experiments) | 200 | 30 ms | 37 ms | 33.5 ms

</div></b>